action_learning_coach/requirements.txt
# Core Dependencies
pyautogen>=0.2.3
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
